+++ public/js/_cache.js
@@ -1,0 +1,4 @@
+if ('serviceWorker' in navigator) {
+  navigator.serviceWorker.register('/sw_cache.js')
+    .catch(console.error);
+}