+++ public/js/_main.js
@@ -1,0 +1,2 @@
+generateButtons();
+window.addEventListener('resize', onResize);