+++ public/js/_predefined.js
@@ -1,0 +1,9 @@
+const audioLinks = [
+  'audio/1.mp3',
+  'audio/2.mp3',
+  'audio/3.mp3',
+  'audio/4.mp3',
+  'audio/5.mp3',
+  'audio/6.mp3',
+  'audio/7.mp3',
+];