+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- build task for gulp
+- watch task for gulp
+- svg template for button
+- adding btn functionality and styles
+- add favicon, manifest, logo images to cache
 
 ## [0.4.0] - 2020-11-10
 ### Added